pyproject.toml
[tool.poetry]
name = "claid"
version = "0.1.0"
description = "AI Song Conversion Community Website"
authors = ["Lee Juneyoung <wnsdud318@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.5"
djangorestframework-simplejwt = "5.2.2"
django = "4.2.1"
djangorestframework = "3.14.0"
django-cors-headers = "4.0.0"
mysqlclient = "2.1.1"
numpy = "1.24.3"
pillow = "9.5.0"
sqlparse = "0.4.4"
django-cleanup = "7.0.0"
django-dotenv = "^1.4.2"
google-api-python-client = "^2.88.0"
google-auth = "^2.19.1"
google-auth-oauthlib = "^1.0.0"
google-auth-httplib2 = "^0.1.0"
requests = "^2.31.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[settings]
virtualenvs.create = true
virtualenvs.in-project = true